# docs/api/api.yaml
...
openapi: 3.0.3
info:
  title: Shelfinity Library Management System API
  description: >
    RESTful API for managing library books, users, borrow/return requests,
    reservations, overdue tracking, email configuration, and reports.
    Authentication is delegated entirely to Keycloak (OIDC) — this API validates
    bearer tokens, it does not issue them or store credentials.
  version: 1.0.0
  contact:
    name: Shadow-Codex
servers:
  - url: http://localhost:9080/api
    description: Local development server
...
paths:
...
  /users:
    post:
      tags: [Users]
      summary: Create a user record
      description: >
        SPEC.md §10.1/§10.3 (both resolved): a non-admin caller may only create a
        record for their own keycloakId, with a caller-supplied role of ADMIN only
        honored when the caller is already an admin; a self-service caller is
        created with `active=false` plus a USER_REGISTRATION queue item, while an
        admin-created account is `active=true` immediately.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUserRequest' }
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserResponse' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: User already exists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    get:
      tags: [Users]
      summary: List all users (admin only)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Users retrieved
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/UserResponse' }
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/me:
    get:
      tags: [Users]
      summary: Get the current user's profile
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Profile retrieved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserResponse' }
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/{id}:
    parameters:
      - $ref: '#/components/parameters/UuidId'
    get:
      tags: [Users]
      summary: Get a user by ID
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: User retrieved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserResponse' }
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Users]
      summary: Update a user, including role reassignment (admin only)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUserRequest' }
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserResponse' }
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Users]
      summary: Delete a user (admin only)
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: User deleted }
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------- Books
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak-issued OIDC access token.
  parameters:
    UuidId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Authenticated but not permitted (e.g. admin-only endpoint)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
    UserRole:
      type: string
      enum: [USER, ADMIN]
...
    UserResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        keycloakId: { type: string }
        email: { type: string, format: email }
        name: { type: string }
        role: { $ref: '#/components/schemas/UserRole' }
        active:
          type: boolean
          description: "false while a USER_REGISTRATION queue item is pending — see SPEC.md §6.1"
        createdAt: { type: string, format: date-time }
    CreateUserRequest:
      type: object
      required: [keycloakId, email, name]
      properties:
        keycloakId: { type: string, example: "123e4567-e89b-12d3-a456-426614174000" }
        email: { type: string, format: email }
        name: { type: string }
        role:
          $ref: '#/components/schemas/UserRole'
          description: "Ignored/forced to USER unless the caller is already an admin — enforced, see SPEC.md §10.1 (resolved)"
...